Turbopack: rename simple tree shaking
